+++ open-metadata-conformance-suite/open-metadata-conformance-suite-server/src/main/java/org/odpi/openmetadata/conformance/tests/performance/delete/TestRelationshipDelete.java
@@ -67,10 +67,12 @@ public class TestRelationshipDelete extends OpenMetadataPerformanceTestCase
         OMRSMetadataCollection metadataCollection = super.getMetadataCollection();
         int numInstances = super.getInstancesPerType();
 
-        Set<String> relationshipsToDelete = getKeys(metadataCollection, numInstances);
-        deleteRelationships(metadataCollection, relationshipsToDelete);
-
-        super.setSuccessMessage("Relationship delete performance tests complete for: " + testTypeName);
+        List<String> methodsToSkip = performanceWorkPad.getMethodsToSkip();
+        if (!methodsToSkip.contains("deleteRelationship")) {
+            Set<String> relationshipsToDelete = getKeys(metadataCollection, numInstances);
+            deleteRelationships(metadataCollection, relationshipsToDelete);
+            super.setSuccessMessage("Relationship delete performance tests complete for: " + testTypeName);
+        }
     }
 
     /**
